organize certificate mock functions
