+++ apps/webcert/src/faker/certificate/fakeCertificate.ts
@@ -1,7 +1,7 @@
 import { PartialDeep } from 'type-fest'
-import { Certificate, CertificateData } from '../types'
+import { Certificate, CertificateData } from '../../types'
+import { fakeResourceLink } from '../fakeResourceLink'
 import { fakeCertificateMetaData } from './fakeCertificateMetaData'
-import { fakeResourceLink } from './fakeResourceLink'
 
 export const fakeCertificate = (data?: PartialDeep<Certificate> & { data?: CertificateData }): Certificate => ({
   metadata: fakeCertificateMetaData(data?.metadata),
