+++ apps/webcert/src/faker/index.ts
@@ -1,10 +1,10 @@
+export * from './certificate/fakeCertificate'
+export * from './certificate/fakeCertificateConfig'
+export * from './certificate/fakeCertificateData'
+export * from './certificate/fakeCertificateDataValidation'
+export * from './certificate/fakeCertificateMetaData'
+export * from './certificate/fakeCertificateValue'
 export * from './fakeCareProvider'
-export * from './fakeCertificate'
-export * from './fakeCertificateConfig'
-export * from './fakeCertificateData'
-export * from './fakeCertificateDataValidation'
-export * from './fakeCertificateMetaData'
-export * from './fakeCertificateValue'
 export * from './fakeCity'
 export * from './fakeDynamicLink'
 export * from './fakeIcf'
